refactors minor for data type

+++ tankwarserver/src/main/java/server/model/dto/Bullet.java
@@ -15,5 +15,5 @@ public class Bullet {
     private Integer positionX;
     private Integer positionY;
     private FaceOrientation faceOrientation;
-    private Float movementSpeed;
+    private Double movementSpeed;
 }
